file upload & download

+++ 17-file_download/model/fileFunc.js
@@ -3,11 +3,16 @@ const path = require('path');
 const fs = require('fs');
 
 function deleteFile(fileUrl) {
+    if(fileUrl.startsWith("http")){
+        console.log("external link");
+        return;
+    }
     const oldFileName = path.basename(fileUrl);
     const oldFilePath = path.join(__dirname, '..', 'uploads', oldFileName);
     // fs.unlink(oldFilePath, (err) => {
     //     if(err) throw err;
     // });
+    console.log(oldFileName);
     return fs.promises.unlink(oldFilePath); // makeing it async
 };
